[issue-117] Rename year_ranges keys 'before/'after' to 'until'/'from' for clarity
Renames the existing :year_ranges keys :before and :after to
:until and :from, respectively. This is done so that the
desired behavior of inclusivity is more clearly specified.

Example: a holiday should be valid up to and including 2002:
  Old syntax - 'before: 2002'
  New syntax - 'until: 2002

This is a breaking change and will require a major semver bump.
All consumers will be required to update to consume these latest
definitions or they will see errors.